fix gameAtIndex, redeploy, multi deployer
