+++ package.json
@@ -31,6 +31,7 @@
     "test-fast": "bun run check && bun run lint && bun run forge && bun run test-components test/gateway/op.test.ts test/gateway/reverse-op.test.ts",
     "lint": "eslint . --ext .ts",
     "build": "bun scripts/build-dist.ts",
+    "deploy": "bun scripts/deploy.ts",
     "loc": "rg -c '^' -g 'src/**/*.ts' -g 'test/**/*.ts' --no-filename | jq -s 'add'"
   },
   "sideEffects": false,
@@ -45,7 +46,7 @@
     "contracts/"
   ],
   "devDependencies": {
-    "@adraffy/blocksmith": "^0.1.56",
+    "@adraffy/blocksmith": "^0.1.58",
     "@eslint/eslintrc": "^3.3.1",
     "@eslint/js": "^9.36.0",
     "@types/bun": "latest",
@@ -54,19 +55,19 @@
     "eslint": "^9.35.0",
     "eslint-config-prettier": "^10.1.8",
     "eslint-plugin-prettier": "^5.5.4",
-    "prettier": "^3.6.2"
-  },
-  "dependencies": {
-    "@namestone/ezccip": "^0.1.1",
+    "prettier": "^3.6.2",
     "@semantic-release/changelog": "^6.0.3",
     "@semantic-release/commit-analyzer": "^13.0.1",
     "@semantic-release/git": "^10.0.1",
     "@semantic-release/github": "^12.0.2",
     "@semantic-release/npm": "^13.1.2",
     "@semantic-release/release-notes-generator": "^14.1.0",
-    "ethers": "^6.15.0",
-    "fzstd": "^0.1.1",
     "semantic-release": "^25.0.2"
+  },
+  "dependencies": {
+    "@namestone/ezccip": "^0.1.1",
+    "ethers": "^6.15.0",
+    "fzstd": "^0.1.1"
   },
   "peerDependencies": {
     "typescript": "^5.6.3"
