+++ src/op/OPFaultRollup.ts
@@ -55,8 +55,8 @@ type ABIFoundGame = {
   rootClaim: string;
 };
 
-const FINDER_MAINNET = '0xDa9883a512b8E1F48bF414f702338F4fAe87D8E8'; // 20251202
-const FINDER_SEPOLIA = '0x76a833f4BF63d843A1Cd12003382066F69699f4d'; // 20251202
+const FINDER_MAINNET = '0x981829755e38b16E6dc23762c1289CA3F4065682'; // 20251205
+const FINDER_SEPOLIA = '0x8CB499cbfe991cAf2d6F18CcCC408dC8265eBE60'; // 20251205
 
 export class OPFaultRollup extends AbstractOPRollup<OPFaultCommit> {
   static readonly ANCHOR_STATE_REGISTRY_ABI = ANCHOR_STATE_REGISTRY_ABI;
